backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
motor==3.6.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
